refactor(chat): extract group and home composer constants
